Split Panner.js
