+++ lib/PannerNode.js
@@ -4,13 +4,8 @@ var AudioNode = require('./AudioNode')
   , FloatPoint3D = require('./FloatPoint3D')
   , DistanceEffect = require('./DistanceEffect')
   , ConeEffect = require('./ConeEffect')
-  , Panner = require('./Panner')
+  , PannerProvider = require('./PannerProvider')
   , mathUtil = require('./mathUtil')
-
-var PanningModelType = {
-  equalpower : true,
-  HRTF       : true,
-}
 
 class PannerNode extends AudioNode {
 
@@ -31,16 +26,13 @@ class PannerNode extends AudioNode {
       },
     })
 
+    /** @type {PannerProvider} */
+    this.pannerProvider = new PannerProvider()
+
     /** @type {PanningModelType} */
-    var _panningModel = 'equalpower'
     Object.defineProperty(this, 'panningModel', {
-      get() { return _panningModel },
-      set(val) {
-        if (!PanningModelType[val]) {
-          throw new Error('Invalid panningModel')
-        }
-        _panningModel = val
-      },
+      get() { return this._pannerProvider.panningModel },
+      set(val) { this._pannerProvider.panningModel = val },
     })
 
     /** @type {float} */
